requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
torchaudio==2.8.0+cu126
transformers==4.56.1
pandas==2.3.2
scikit-learn==1.7.2
PyYAML==6.0.2
datasets==4.0.0
pyarrow==21.0.0